feat: add TodoApp component
